+++ src/main.rs
@@ -9,7 +9,7 @@ fn main() {
         Builder::from_env(Env::default().default_filter_or("info")).init();
     }
 
-    if let Err(e) = ics_dm_iot_module_rs::run() {
+    if let Err(e) = iot_module_template_rs::run() {
         error!("Application error: {}", e);
 
         process::exit(1);